Kamera fungerar
