--- CyberPunkRPG - prototyp/CyberPunkRPG.sln
@@ -1,22 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Express 14 for Windows Desktop
-VisualStudioVersion = 14.0.25420.1
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CyberPunkRPG", "CyberPunkRPG\CyberPunkRPG.csproj", "{5254D774-F823-4A65-BA4B-0F4310CE8CFF}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|x86 = Debug|x86
-		Release|x86 = Release|x86
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{5254D774-F823-4A65-BA4B-0F4310CE8CFF}.Debug|x86.ActiveCfg = Debug|x86
-		{5254D774-F823-4A65-BA4B-0F4310CE8CFF}.Debug|x86.Build.0 = Debug|x86
-		{5254D774-F823-4A65-BA4B-0F4310CE8CFF}.Release|x86.ActiveCfg = Release|x86
-		{5254D774-F823-4A65-BA4B-0F4310CE8CFF}.Release|x86.Build.0 = Release|x86
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-EndGlobal
