--- CyberPunkRPG - prototyp/CyberPunkRPG/Constants.cs
@@ -1,14 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace CyberPunkRPG
-{
-    class Constants
-    {
-        public static int ScreenWidth = 1920;
-        public static int ScreenHeight = 1080;
-    }
-}
