--- CyberPunkRPG - prototyp/CyberPunkRPG/Program.cs
@@ -1,22 +1,0 @@
-﻿using System;
-
-namespace CyberPunkRPG
-{
-#if WINDOWS || LINUX
-    /// <summary>
-    /// The main class.
-    /// </summary>
-    public static class Program
-    {
-        /// <summary>
-        /// The main entry point for the application.
-        /// </summary>
-        [STAThread]
-        static void Main()
-        {
-            using (var game = new Game1())
-                game.Run();
-        }
-    }
-#endif
-}
