+++ CyberPunkRPG/CyberPunkRPG/Camera.cs
@@ -1,4 +1,6 @@
-﻿using System;
+﻿using Microsoft.Xna.Framework;
+using Microsoft.Xna.Framework.Graphics;
+using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
@@ -8,5 +10,34 @@ namespace CyberPunkRPG
 {
     class Camera
     {
+        public Vector2 position;
+        public Viewport view;
+        public Matrix transform;
+
+        public Camera(Viewport view)
+        {
+            this.view = view;
+        }
+
+        public void SetPosition(Vector2 position)
+        {
+            this.position = position;
+        }
+
+        public Vector2 GetPosition()
+        {
+            return position;
+        }
+
+        public Matrix ReturnTransform()
+        {
+            return transform;
+        }
+
+        public Matrix GetTransformation(Viewport view)
+        {
+            transform = Matrix.CreateTranslation(-position.X + view.Width / 2, -position.Y + view.Height / 2, 0);
+            return transform;
+        }
     }
 }
